requirements.txt
fastapi
uvicorn
scikit-learn
spacy
PyPDF2
python-docx
pandas
numpy
pytest
python-multipart
jinja2
